refactor(170): reorganize domain puzzle assignment tests
